- ajustado unitary_value para decimal

+++ vc-module-MelhorEnvio/vc_module_MelhorEnvio.Core/Models/CartOut.cs
@@ -16,7 +16,7 @@ namespace vc_module_MelhorEnvio.Core.Models
             public int Quantity { get; set; }
 
             [JsonProperty("unitary_value")]
-            public int UnitaryValue { get; set; }
+            public decimal UnitaryValue { get; set; }
 
             [JsonProperty("weight")]
             public object Weight { get; set; }
@@ -80,16 +80,16 @@ namespace vc_module_MelhorEnvio.Core.Models
         public object ServiceCode { get; set; }
 
         [JsonProperty("quote")]
-        public double Quote { get; set; }
+        public decimal Quote { get; set; }
 
         [JsonProperty("price")]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [JsonProperty("coupon")]
         public object Coupon { get; set; }
 
         [JsonProperty("discount")]
-        public double Discount { get; set; }
+        public decimal Discount { get; set; }
 
         [JsonProperty("delivery_min")]
         public int DeliveryMin { get; set; }
@@ -104,7 +104,7 @@ namespace vc_module_MelhorEnvio.Core.Models
         public object Reminder { get; set; }
 
         [JsonProperty("insurance_value")]
-        public double InsuranceValue { get; set; }
+        public decimal InsuranceValue { get; set; }
 
         [JsonProperty("weight")]
         public object Weight { get; set; }
